requirements.txt
streamlit
moviepy==1.0.3
edge-tts
langchain-groq
requests
python-dotenv
imageio-ffmpeg
playwright